+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod admin;
 pub mod general;
 pub mod hooks;
 pub mod minecraft;